+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 0.16.18
 -------
 - Support custom function in update. (#537)
+- Add `Model.refresh_from_db`. (#549)
 
 0.16.17
 -------